requirements.txt
async-generator==1.10
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
digital-sport==0.1.2
h11==0.13.0
idna==3.3
MouseInfo==0.1.3
outcome==1.1.0
Pillow==9.1.0
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyOpenSSL==22.0.0
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
python3-xlib==0.15
pytweening==1.0.4
requests==2.27.1
selenium==4.1.3
sniffio==1.2.0
sortedcontainers==2.4.0
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.9
webdriver-manager==3.5.4
wsproto==1.1.0
